[WEBDEV-761] Log to stdout apart from in development
This change removes all disk logging and replaces it with just logging to stdout.
Access logs are the same as before with an added key of "type" of which the value is "access".

When in production environments (ie. `NODE_ENV` is `'production'` or `PRODUCTION_LOGGING` is `'true'`), Shunter logs will be in a JSON format. Otherwise, it will be in a friendly format with colours.

Other changes:
- Removed volume declaration from Dockerfile
- Removed winston transports and filters that handled logging to the disk
- Added `timekeeper` for time-dependent tests
- Added tests for winston-console transport